updated distance return

+++ src/main/java/de/wirvsvirus/hack/rest/model/ProviderResponseModel.java
@@ -44,7 +44,7 @@ public class ProviderResponseModel {
 
 	private boolean overnightPossible;
 
-	private float distance;
+	private double distance;
 
 	private double latitude;
